+++ src/main/scala/org/broadinstitute/hail/driver/ImportAnnotations.scala
@@ -6,7 +6,6 @@ import org.broadinstitute.hail.annotations.Annotation
 import org.broadinstitute.hail.expr._
 import org.broadinstitute.hail.utils.{TextTableConfiguration, TextTableReader}
 import org.broadinstitute.hail.variant._
-import org.json4s.jackson.JsonMethods._
 import org.kohsuke.args4j.{Argument, Option => Args4jOption}
 
 import scala.collection.JavaConverters._
@@ -17,7 +16,6 @@ object ImportAnnotations extends SuperCommand {
   def description = "Import variants and annotations as a sites-only VDS"
 
   register(ImportAnnotationsTable)
-  register(ImportAnnotationsJSON)
 }
 
 object ImportAnnotationsTable extends Command {
@@ -120,62 +118,4 @@ object ImportAnnotationsTable extends Command {
     state.copy(vds = vds)
   }
 
-}
-
-object ImportAnnotationsJSON extends Command {
-
-  class Options extends BaseOptions {
-    @Argument(usage = "<files...>")
-    var arguments: java.util.ArrayList[String] = new java.util.ArrayList[String]()
-
-    @Args4jOption(required = true, name = "-t", aliases = Array("--type"),
-      usage = "Type of imported JSON")
-    var `type`: String = _
-
-    @Args4jOption(required = true, name = "-e", aliases = Array("--vfields"),
-      usage = "Expressions for chromosome, position, ref and alt in terms of `root'")
-    var variantFields: String = _
-  }
-
-  def newOptions = new Options
-
-  def name = "importannotations json"
-
-  def description = "Import variants and annotations from JSON as a sites-only VDS"
-
-  def requiresVDS = false
-
-  def supportsMultiallelic = true
-
-  def run(state: State, options: Options): State = {
-    val sc = state.sc
-
-    val files = hadoopGlobAll(options.arguments.asScala, state.hadoopConf)
-
-    if (files.isEmpty)
-      fatal("Arguments referred to no files")
-
-    val t = Parser.parseType(options.`type`)
-
-    val extractVariant = JSONAnnotationImpex.jsonExtractVariant(t, options.variantFields)
-
-    val rdd =
-      sc.union(files.map { f =>
-        sc.textFile(f)
-          .map { line =>
-            JSONAnnotationImpex.importAnnotation(parse(line), t, "<root>")
-          }
-      })
-        .flatMap { va =>
-          extractVariant(va)
-            .map { v => (v, va, Iterable.empty[Genotype]) }
-        }
-
-    val vds = new VariantDataset(
-      VariantMetadata(IndexedSeq.empty, Annotation.emptyIndexedSeq(0), Annotation.empty,
-        TStruct.empty, t, TStruct.empty, wasSplit = true),
-      rdd)
-
-    state.copy(vds = vds)
-  }
 }